Separated desktop WindowFocusListener class
This is actually so the Android code wouldn't fail. Since I added
WindowFocusListener before, the Android compiler refused to dex this
class, and when Dependencies tried to register it initially, the Android
app was throwing a ClassNotFoundException on startup. Still need to redo
control flow so that desktop code is never even referenced from Android
code.