Now using InputStream for custom fonts
Now using InputStream instead of File for creating custom Font objects.
On Desktop, this means no temp file needs to be created (it was not
possible to easily delete the temp file on shutdown). On Android,
nothing changes - a temp file is still needed (which is fine - the temp
file was being deleted okay on shutdown anyway).