Redefined measurements in FontLayout
Now using the max ascent and max descent ("top" and "bottom" in Android)
for FontLayout and moved calculation for FontLayout#height up into
superclass. Now standardized on using all positive values for ascent
(Android uses a negative value, since it is relative to baseline, but
Desktop uses a positive value, since it is the absolute difference).